Introduce lpmd_state_machine.c
Introduce lpmd_state_machine.c and move all state machine related code
from lpmd_proc.c to lpmd_state_machine.c.
