+++ Makefile.am
@@ -58,6 +58,7 @@ intel_lpmd_SOURCES = \
 	src/lpmd_irq.c \
 	src/lpmd_socket.c \
 	src/lpmd_util.c \
+	src/lpmd_state_machine.c \
 	src/wlt_proxy/wlt_proxy.c \
 	src/wlt_proxy/spike_mgmt.c \
 	src/wlt_proxy/state_machine.c \
